refactor(server): split lib.rs into focused modules
Split the 1116-line lib.rs into four modules:
- push_constants.rs: all push constant structs
- passes.rs: ComputePass struct, create_pass(), dispatch(), barrier()
- simulation.rs: GpuSimulation impl (new, step, render, destroy)
- readback.rs: readback and buffer accessor methods

lib.rs now contains only mod declarations, re-exports, ServerError, and tests.
Public API is preserved via re-exports.

Closes #185
