+++ crates/server/src/readback.rs
@@ -1,0 +1,57 @@
+//! GPU readback and buffer accessor methods for [`GpuSimulation`].
+
+use ash::vk;
+use gpu_core::{
+    VulkanContext,
+    buffer::{self, GpuBuffer},
+};
+use shared::Particle;
+
+use crate::{Result, simulation::GpuSimulation};
+
+impl GpuSimulation {
+    /// Read particle data back from the GPU for debugging.
+    ///
+    /// This is a synchronous operation that stalls the GPU pipeline.
+    /// Use only for debugging and testing.
+    pub fn readback_particles(&self, ctx: &VulkanContext) -> Result<Vec<Particle>> {
+        let particles =
+            buffer::readback::<Particle>(ctx, &self.particle_buffer, self.num_particles as usize)?;
+        Ok(particles)
+    }
+
+    /// Returns the current number of active particles.
+    pub fn num_particles(&self) -> u32 {
+        self.num_particles
+    }
+
+    /// Returns a reference to the particle buffer.
+    pub fn particle_buffer(&self) -> &GpuBuffer {
+        &self.particle_buffer
+    }
+
+    /// Returns a reference to the grid buffer.
+    pub fn grid_buffer(&self) -> &GpuBuffer {
+        &self.grid_buffer
+    }
+
+    /// Returns a reference to the voxel buffer.
+    pub fn voxel_buffer(&self) -> &GpuBuffer {
+        &self.voxel_buffer
+    }
+
+    /// Returns the Vulkan buffer handle for the render output.
+    ///
+    /// The buffer contains packed BGRA u32 pixels, sized `width * height`.
+    pub fn render_output_buffer(&self) -> vk::Buffer {
+        self.render_output_buffer.buffer
+    }
+
+    /// Returns a reference to the render output [`GpuBuffer`].
+    ///
+    /// Useful for readback operations (e.g., headless screenshot).
+    /// The buffer contains packed BGRA u32 pixels, sized `width * height`.
+    pub fn render_output_gpu_buffer(&self) -> &GpuBuffer {
+        &self.render_output_buffer
+    }
+}
